+++ endpoint/fetcher.py
@@ -1,16 +1,21 @@
-import logging
-import requests
+from core.core_log import get_logger
 import urllib.parse
 from datetime import datetime
-from bs4 import BeautifulSoup as bs
-from selenium import webdriver
-from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.common.by import By
-from selenium.webdriver.chrome.options import Options
-from selenium.common.exceptions import WebDriverException
-from pyvirtualdisplay import Display
 
-log = logging.getLogger("api.fetcher")
+import requests
+from bs4 import BeautifulSoup as bs
+from pyvirtualdisplay import Display
+from selenium import webdriver
+from selenium.common.exceptions import WebDriverException
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.common.by import By
+
+log = get_logger(__name__)
+
+
+# ----------------------------------------------------------------------------------------------------------------
+# The fetching module. Contains data fetcher for retrieving data from defined endpoints.
+# ----------------------------------------------------------------------------------------------------------------
 
 class UlmDeFetcher:
 
@@ -21,7 +26,7 @@ class UlmDeFetcher:
         """
         try:
             # set connect and read timeout to 5 seconds
-            response = requests.get('https://www.ulm.de/', timeout=(5,5))
+            response = requests.get('https://www.ulm.de/', timeout=(5, 5))
         except (requests.exceptions.Timeout, requests.exceptions.ConnectionError) as e:
             log.error(f"Ulm.de connection problem: " + str(e))
             return None
@@ -60,7 +65,7 @@ class WetterComFetcher:
         """
         try:
             # set connect and read timeout to 5 seconds
-            response = requests.get(url, timeout=(5,5))
+            response = requests.get(url, timeout=(5, 5))
         except (requests.exceptions.Timeout, requests.exceptions.ConnectionError) as e:
             log.error(f"Wetter.com connection problem: " + str(e))
             return None
@@ -78,7 +83,7 @@ class WetterComFetcher:
                 log.error("Temperature element not found on the page.")
         else:
             log.error("Failed to retrieve weather data.")
-        
+
         return None
 
     @staticmethod
@@ -94,7 +99,7 @@ class WetterComFetcher:
         service = webdriver.ChromeService(executable_path='/usr/lib/chromium-browser/chromedriver')
         driver = webdriver.Chrome(service=service, options=options)
         timeout_s = 30
-        driver.set_page_load_timeout(timeout_s) 
+        driver.set_page_load_timeout(timeout_s)
         driver.implicitly_wait(timeout_s)
         try:
             driver.get(url)
@@ -119,7 +124,7 @@ class GoogleFetcher:
         :location: e.g. "New York"
         """
 
-        url = "https://www.google.com/search?lr=lang_en&ie=UTF-8&q=weather" + location
+        url = "https://www.google.com/search?lr=lang_en&ie=UTF-8&q=weather%20" + location
         user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/94.0.4606.81 Safari/537.36"
         language = "en-US,en;q=0.5"
         try:
@@ -128,7 +133,7 @@ class GoogleFetcher:
             session.headers["Accept-Language"] = language
             session.headers["Content-Language"] = language
             # set connect and read timeout to 5 seconds
-            html = session.get(url, timeout=(5,5))
+            html = session.get(url, timeout=(5, 5))
             soup = bs(html.text, "html.parser")
 
             return {"region": soup.find("div", attrs={"id": "wob_loc"}).text,
@@ -136,7 +141,9 @@ class GoogleFetcher:
                     "precipitation": float(soup.find("span", attrs={"id": "wob_pp"}).text.replace("%", "")),
                     "humidity": float(soup.find("span", attrs={"id": "wob_hm"}).text.replace("%", "")),
                     "wind": float(soup.find("span", attrs={"id": "wob_ws"}).text.replace(" km/h", ""))}
-        except (WebDriverException, AttributeError, requests.exceptions.Timeout, requests.exceptions.ConnectionError) as e:
+        except (
+                WebDriverException, AttributeError, requests.exceptions.Timeout,
+                requests.exceptions.ConnectionError) as e:
             log.error("Google Weather connection problem: " + str(e))
             return None
 
@@ -152,7 +159,7 @@ class Fetcher:
         self.endpoint = endpoint
         self.params = None if not params else urllib.parse.urlencode(params)
         self.api_link = self._create_api_link()
-        self.timeouts = (30,5) # connect timeout 30s and read timeout 5s
+        self.timeouts = (30, 5)  # connect timeout 30s and read timeout 5s
 
     def __str__(self):
         return f"Fetcher[{self.api_link}]"