+++ hometemp.py
@@ -1,59 +1,34 @@
-import Adafruit_DHT, configparser, logging, re, schedule, time, threading
-from datetime import datetime, timedelta
-from gpiozero import CPUTemperature
-from distribute.email import EmailDistributor
-from persist.database import DwDDataHandler, GoogleDataHandler, UlmDeHandler, SensorDataHandler, WetterComHandler
-from util.manager import PostgresDockerManager
-from visualize.plots import draw_plots
+from core.core_log import setup_logging, get_logger
+from core.core_configuration import load_config, database_config, hometemp_config
+import re
+import schedule
+import threading
+import time
+from datetime import datetime
 
-# only logs from this file will be written to console but all logs will be saved to file
-for handler in logging.root.handlers[:]:
-    logging.root.removeHandler(handler)
-logging.basicConfig(filename='measurement.log',
-                    encoding='utf-8',
-                    level=logging.INFO)
-config = configparser.ConfigParser()
-config.read('hometemp.ini')
-log = logging.getLogger('hometemp')
+from core.command import CommandService
+from core.distribute import send_visualization_email
+from core.database import DwDDataHandler, GoogleDataHandler, UlmDeHandler, SensorDataHandler, WetterComHandler
+from core.sensors.dht import get_sensor_data
+from core.sensors.util import get_temperature
+from core.virtualization import init_postgres_container
+from core.plotting import draw_plots
 
-
-# ------------------------------- Raspberry Pi Temps ----------------------------------------------
-
-def get_temperature():
-    cpu = CPUTemperature()
-    return float(cpu.temperature)
+# GLOBAL Variables
+log = None
+command_service = None
 
 
-def get_sensor_data():
-    """
-    Returns temperature and humidity data measures by AM2302 Sensor and the measure timestamp or raise exception 
-    if the values could not be retrieved
-    """
-    DHT_SENSOR = Adafruit_DHT.AM2302
-    DHT_PIN = 2
-    humidity, temperature = Adafruit_DHT.read_retry(DHT_SENSOR, DHT_PIN)
-
-    if humidity is not None and temperature is not None:
-        # postgres expects timestamp ins ISO 8601 format
-        timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        return temperature, humidity, timestamp
-    else:
-        log.error("Failed to retrieve data from AM2302 sensor")
-        return None, None, None
-
-
-# ------------------------------- Main  ----------------------------------------------
-
-def create_and_backup_visualization():
-    log.info("Creating Measurement Data Visualization")
-    auth = config["db"]
+def _get__visualization_data():
+    auth = database_config()
     sensor_data_handler = SensorDataHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'],
                                             'sensor_data')
     sensor_data_handler.init_db_connection(check_table=False)
     # sensor data
     df = sensor_data_handler.read_data_into_dataframe()
     df = df.sort_values(by="timestamp")
-    df['timestamp'] = df['timestamp'].map(lambda x: datetime.strptime(str(x).strip(), '%Y-%m-%d %H:%M:%S'))
+    df['timestamp'] = df['timestamp'].map(
+        lambda x: datetime.strptime(str(x).replace("+00:00", "").strip(), '%Y-%m-%d %H:%M:%S'))
     # Google weather data
     google_handler = GoogleDataHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'], 'google_data')
     google_handler.init_db_connection(check_table=False)
@@ -79,11 +54,40 @@ def create_and_backup_visualization():
     ulmde_handler = UlmDeHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'], 'ulmde_data')
     ulmde_handler.init_db_connection()
     ulmde_df = ulmde_handler.read_data_into_dataframe()
-    ulmde_df['timestamp'] = ulmde_df['timestamp'].map(lambda x: datetime.strptime(str(x).strip().replace('+00:00', ''), '%Y-%m-%d %H:%M:%S'))
+    ulmde_df['timestamp'] = ulmde_df['timestamp'].map(
+        lambda x: datetime.strptime(str(x).strip().replace('+00:00', ''), '%Y-%m-%d %H:%M:%S'))
 
-    draw_plots(df, google_df=google_df, dwd_df=dwd_df, wettercom_df=wettercom_df, ulmde_df=ulmde_df)
+    return (df, google_df, dwd_df, wettercom_df, ulmde_df)
+
+
+# ------------------------------- Main  ----------------------------------------------
+
+def _create_visualization_commanded(commander):
+    log.info("Command: Creating Measurement Data Visualization")
+    sensor_data, google_df, dwd_df, wettercom_df, ulmde_df = _get__visualization_data()
+
+    name = datetime.now().strftime("%d-%m-%Y")
+    save_path = f"plots/commanded/{name}.pdf"
+    draw_plots(df=sensor_data, google_df=google_df, dwd_df=dwd_df, wettercom_df=wettercom_df, ulmde_df=ulmde_df,
+               save_path=save_path)
+    log.info("Command: Done")
+    send_visualization_email(df=sensor_data, ulmde_df=ulmde_df, google_df=google_df, dwd_df=dwd_df,
+                             wettercom_df=wettercom_df, path_to_pdf=save_path, receiver=commander)
+
+
+def create_visualization_timed():
+    log.info("Timed: Creating Measurement Data Visualization")
+    sensor_data, google_df, dwd_df, wettercom_df, ulmde_df = _get__visualization_data()
+    draw_plots(df=sensor_data, google_df=google_df, dwd_df=dwd_df, wettercom_df=wettercom_df, ulmde_df=ulmde_df)
+    log.info("Timed: Done")
+    send_visualization_email(df=sensor_data, ulmde_df=ulmde_df, google_df=google_df, dwd_df=dwd_df,
+                             wettercom_df=wettercom_df)
+
+
+def run_received_commands():
+    log.info("Checking for commands")
+    command_service.receive_and_execute_commands()
     log.info("Done")
-    EmailDistributor.send_visualization_email(df, google_df=google_df, dwd_df=dwd_df, wettercom_df=wettercom_df)
 
 
 def run_threaded(job_func):
@@ -93,55 +97,52 @@ def run_threaded(job_func):
 
 def collect_and_save_to_db():
     log.info("Start Measurement Data Collection")
-    auth = config["db"]
+    auth = database_config()
     handler = SensorDataHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'], 'sensor_data')
     handler.init_db_connection()
     cpu_temp = get_temperature()
-    room_temp, humidity, timestamp = get_sensor_data()
+    room_temp, humidity, timestamp = get_sensor_data(int(hometemp_config()["sensor_pin"]), False)
     if room_temp is not None and humidity is not None and timestamp is not None:
         log.info(
-            "[Measurement {0}] CPU={1:f}*C, Room={2:f}*C, Humidity={3:f}%".format(timestamp, cpu_temp, room_temp, humidity))
-        handler.insert_measurements_into_db(timestamp=timestamp, humidity=humidity, room_temp=room_temp, cpu_temp=cpu_temp)
+            "[Measurement {0}] CPU={1:f}*C, Room={2:f}*C, Humidity={3:f}%".format(timestamp, cpu_temp, room_temp,
+                                                                                  humidity))
+        handler.insert_measurements_into_db(timestamp=timestamp, humidity=humidity, room_temp=room_temp,
+                                            cpu_temp=cpu_temp)
     log.info("Done")
 
 
-def init_postgres_container():
-    log.info("Checking for existing database")
-    auth = config["db"]
-    docker_manager = PostgresDockerManager(auth["db_name"], auth["db_user"], auth["db_pw"])
-    container_name = auth["container_name"]
-    if docker_manager.container_exists(container_name):
-        log.info("Reusing existing database container")
-        return docker_manager.start_container(container_name)
-    else:
-        log.info("No database container found. Creating database container")
-        if docker_manager.pull_postgres_image():
-            docker_manager.create_postgres_container(container_name)
-            return docker_manager.start_container(container_name)
-
-    return False
-
-
 def main():
-    log = logging.getLogger('hometemp')
-    log.addHandler(logging.StreamHandler())
-    log.info(f"------------------- HomeTemp v{config['hometemp']['version']} -------------------")
-    if not init_postgres_container():
+    log.info(f"------------------- HomeTemp v{hometemp_config()['version']} -------------------")
+    if not init_postgres_container(database_config()):
         log.error("Postgres container startup error! Shutting down ...")
         exit(1)
         # after restart, database needs some time to start
     time.sleep(1)
+
+    cmd_name = 'plot'
+    function_params = ['commander']
+    command_service.add_new_command((cmd_name, [], _create_visualization_commanded, function_params))
+
     schedule.every(10).minutes.do(collect_and_save_to_db)
     # run_threaded assumes that we never have overlapping usage of this method or its components
-    schedule.every().day.at("06:00").do(run_threaded, create_and_backup_visualization)
+    schedule.every().day.at("06:00").do(run_threaded, create_visualization_timed)
+    schedule.every(10).minutes.do(run_threaded, run_received_commands)
     log.info("finished initialization")
 
     collect_and_save_to_db()
-    run_threaded(create_and_backup_visualization)
+    create_visualization_timed()
+    time.sleep(1)
+    run_received_commands()
+    log.info("entering main loop")
     while True:
         schedule.run_pending()
         time.sleep(1)
 
 
 if __name__ == "__main__":
+    setup_logging(log_file='hometemp.log')
+    load_config()
+    # Define all global variables
+    log = get_logger(__name__)
+    command_service = CommandService()
     main()