+++ persist/persistence_logger.py
@@ -1,4 +1,0 @@
-import logging
-
-logging.basicConfig(level=logging.INFO)
-per_log = logging.getLogger('persistence')