+++ tests/database_test.py
@@ -1,11 +1,10 @@
-from persist.database import SensorDataHandler, DwDDataHandler
-from api.fetcher import DWDFetcher
-import configparser
+from core.core_configuration import database_config, dwd_config
 from datetime import datetime, timedelta
 
-config = configparser.ConfigParser()
-config.read('hometemp.ini')
-auth = config["db"]
+from core.database import SensorDataHandler, DwDDataHandler
+from endpoint.fetcher import DWDFetcher
+
+auth = database_config()
 
 
 def foo():
@@ -29,7 +28,7 @@ def foo():
 
 def process_data_updates():
     sanity_threshold = 100
-    fetcher = DWDFetcher(config["dwd"]["station"])
+    fetcher = DWDFetcher(dwd_config()["station"])
     data = fetcher.get_dwd_data()
     handler = DwDDataHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'], 'dwd_data')
     handler.init_db_connection()