+++ visualize/vis_logger.py
@@ -1,4 +1,0 @@
-import logging
-
-logging.basicConfig(level=logging.INFO)
-vis_log = logging.getLogger('visualize')